Refs #21162: Split example tests in different file / executable tests
