+++ test/examples/test_configuration.py
@@ -1,4 +1,3 @@
-
 import subprocess
 import pytest
 
@@ -153,45 +152,6 @@ def test_configuration_expected_output(pub_args, sub_args, expected_message, n_m
 
     except subprocess.CalledProcessError as e:
         print (render_out)
-    except subprocess.TimeoutExpired:
-        print('TIMEOUT')
-        print(out)
-
-    assert(ret)
-
-def test_hello_world():
-    """."""
-    ret = False
-    out = ''
-    try:
-        out = subprocess.check_output(
-            '@DOCKER_EXECUTABLE@ compose -f hello_world.compose.yml up',
-            stderr=subprocess.STDOUT,
-            shell=True,
-            timeout=30
-        ).decode().split('\n')
-
-        sent = 0
-        received = 0
-        for line in out:
-            if 'SENT' in line:
-                sent += 1
-                continue
-
-            if 'RECEIVED' in line:
-                received += 1
-                continue
-
-        if sent != 0 and received != 0 and sent * 2 == received:
-            ret = True
-        else:
-            print('ERROR: sent: ' + str(sent) + ', but received: ' + str(received) +
-                  ' (expected: ' + str(sent * 2) + ')')
-            raise subprocess.CalledProcessError(1, '')
-
-    except subprocess.CalledProcessError:
-        for l in out:
-            print(l)
     except subprocess.TimeoutExpired:
         print('TIMEOUT')
         print(out)
